Implement user spcific tasks operations

+++ src/tasks/tasks.service.ts
@@ -1,11 +1,12 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 
+import User from '../auth/user.entity';
 import CreateTaskDto from './dto/create-task.dto';
 import FilterTasksDto from './dto/filter-tasks.dto';
-import TaskRepository from './task.repository';
-import Task from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import Task from './task.entity';
+import TaskRepository from './task.repository';
 
 @Injectable()
 export default class TasksService {
@@ -14,15 +15,18 @@ export default class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne({ id });
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
     if (!found) {
       throw new NotFoundException(`Task With ID ${id} isn't found`);
     }
+    delete found.userId;
     return found;
   }
 
-  async getTasks(filterDto: FilterTasksDto): Promise<Task[]> {
+  async getTasks(filterDto: FilterTasksDto, user: User): Promise<Task[]> {
     const { status, search } = filterDto;
     const query = this.taskRepository.createQueryBuilder('task');
     if (status) {
@@ -35,23 +39,28 @@ export default class TasksService {
         { search: `%${search}%` },
       );
     }
+    query.where('task.userId = :userId', { userId: user.id });
     const tasks = await query.getMany();
     return tasks;
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: string): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`Task With ID ${id} isn't found`);
     }
   }
 
-  async updateTaskTypeById(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskTypeById(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
